Fix home.index blade
